main qml: reset state on test start, not on click
Tests can be started via means other than clicks.  Namely, via shorting
out a GPIO.  That means we'll be looking for signals to indicate tests
starting, instead of just looking for mouse clicks.

Alter the way the UI is reset.  Reset the UI when the test is started, not
when a click is handled.  This will allow us to handle the case where the
test is started by a means other than a click.

Signed-off-by: Sean Cross <sean@xobs.io>